Tiny fix of models

+++ Dto/VehicleDto.cs
@@ -7,8 +7,8 @@ namespace Autopark.Dto
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public double Price { get; set; }
-        public double ZeroToHundred { get; set; }
+        public int Price { get; set; }
+        public float ZeroToHundred { get; set; }
         public int Mileage { get; set; }
         public int Year { get; set; }
         public int HorsePower { get; set; }
